+++ src/CognitiveServices/CognitiveServices/ChangeLog.md
@@ -19,6 +19,8 @@
 -->
 ## Upcoming Release
 
+* Support Identity, Encryption, UserOwnedStorage 
+
 ## Version 1.2.3
 * Updated SDK to 7.0
 * Improved error message when server responses empty body